catalogue/bpf: Replace usage of sync with atomic_fetch
BPF doesn't have a real sync instruction so we were using an
artificially created sync instruction to provide ordering. Now herd
supports atomic operations with fetch and they can be used to enforce
full ordering.

Replace all usages fo sync with
r6 = atomic_fetch_add((u64*)(r5 + 0), r6)

this is just adding 0 to a variable that is already 0. It is a NOP but
provides full ordering.

Signed-off-by: Puranjay Mohan <puranjay@kernel.org>